% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_counts.R
\name{compute_signatures_for_all_examples}
\alias{compute_signatures_for_all_examples}
\title{\code{compute_signatures_for_all_examples} Compute mutational signatures and \cr
find changepoints}
\usage{
save_data_for_samples(dir_counts = TrackSig.options()$DIR_COUNTS,
  bootstrap_counts = TrackSig.options()$BOOTSTRAP_COUNTS)

compute_signatures_for_all_examples(countsDir, bootstrapDir,
  samples_to_run = c())


  compute_errorbars_for_all_examples(bootstrap_counts = TrackSig.options()$BOOTSTRAP_COUNTS)

compute_signatures_for_all_examples(countsDir, bootstrapDir,
  samples_to_run = c())
}
\arguments{
\item{dir_counts}{Directory where previously computed mutation counts are \cr
located. Defaults to TrackSig.options()$DIR_COUNTS}

\item{bootstrap_counts}{Directory where previously computed mutation bootstrapped\cr
counts are located. Defaults toTrackSig.options()$BOOTSTRAP_COUNTS}
}
\description{
\code{compute_signatures_for_all_examples} Compute mutational signatures and \cr
find changepoints
}
